Functionality to put schedules, code cleanup

+++ src/main/java/com/xyz/platformsvc/helper/ShowScheduleOpsHelper.java
@@ -7,25 +7,31 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import org.springframework.util.CollectionUtils;
 
 import com.xyz.dal.entity.theater.show.ShowEntity;
+import com.xyz.dal.entity.theater.show.ShowPricingEntity;
 import com.xyz.dal.entity.theater.show.ShowScheduleEntity;
-import com.xyz.dal.repository.ShowRepository;
-import com.xyz.dal.repository.ShowScheduleRepository;
-import com.xyz.dal.repository.TheaterMovieCatalogRepository;
-import com.xyz.dal.repository.TheaterScreenRepository;
+import com.xyz.dal.repository.show.ShowRepository;
+import com.xyz.dal.repository.show.ShowScheduleRepository;
+import com.xyz.dal.repository.theater.TheaterMovieCatalogRepository;
+import com.xyz.dal.repository.theater.screen.TheaterScreenRepository;
+import com.xyz.platformsvc.exception.InvalidRequestException;
+import com.xyz.platformsvc.exception.PlatformServiceException;
+import com.xyz.platformsvc.exception.ResourceNotFoundException;
 import com.xyz.platformsvc.mapper.ShowScheduleMapper;
 import com.xyz.platformsvc.rest.model.show.ShowSchedule;
 
 @Component
 public class ShowScheduleOpsHelper {
-
+	
+	private static final Logger logger = LoggerFactory.getLogger(ShowScheduleOpsHelper.class);
+	
 	@Autowired
 	ShowScheduleMapper showScheduleMapper;
 	
@@ -45,38 +51,49 @@ public class ShowScheduleOpsHelper {
 		return showScheduleMapper.toRestObj(showScheduleEntity);
 	}
 	
-	public ShowSchedule createShowSchedule(ShowSchedule showSchedule) {
+	public ShowSchedule createShowSchedule(ShowSchedule showSchedule) throws PlatformServiceException {
 		ShowScheduleEntity showScheduleEntity = showScheduleMapper.toEntityObj(showSchedule);
-		showScheduleEntity = showScheduleRepository.save(showScheduleEntity);
+		try {
+			showScheduleEntity = showScheduleRepository.saveAndFlush(showScheduleEntity);
+		} catch(Exception e) {
+			String errorString = "Fail to create show schedule. Exception: "+e.getMessage();
+			logger.error(errorString);
+			throw new PlatformServiceException(errorString, e);
+		}
 		return  getShow(showScheduleEntity);
 	}
 
 	public List<ShowSchedule> getShowSchedule(Long theaterId, Long movieId, LocalDate date) {
 		List<ShowScheduleEntity> showScheduleList = showScheduleRepository.lookupShowSchedule(theaterId, movieId, date);
-		if(CollectionUtils.isEmpty(showScheduleList)) {
-			//exception
-		}
-		
 		return showScheduleList.stream().map(showScheduleMapper::toRestObj).collect(Collectors.toList());
 	}
 	
-	public ShowSchedule getShowSchedule(Long showScheduleId) {
+	public ShowSchedule getShowSchedule(Long showScheduleId) throws ResourceNotFoundException {
 		ShowScheduleEntity showScheduleEntity = showScheduleRepository.lookupShowSchedule(showScheduleId);
-		
+		if(showScheduleEntity==null) {
+			String errorString ="Fail to find schedule having id: "+showScheduleId;
+			throw new ResourceNotFoundException(errorString);
+		}
 		return showScheduleMapper.toRestObj(showScheduleEntity);
 	}
 
-	public ShowSchedule updateShowSchedule(Long showScheduleId, ShowSchedule showSchedule) {
+	public ShowSchedule updateShowSchedule(Long showScheduleId, ShowSchedule showSchedule) throws ResourceNotFoundException, InvalidRequestException, PlatformServiceException {
 		//lookup existing schedule 
 		final ShowScheduleEntity existingEntity = showScheduleRepository.lookupShowSchedule(showScheduleId);
-		
+		if(existingEntity==null) {
+			String errorString ="Fail to find schedule having id: "+showScheduleId;
+			throw new ResourceNotFoundException(errorString);
+		}
+
 		//new 
 		ShowScheduleEntity newEntity = showScheduleMapper.toEntityObj(showSchedule);
 		
 		if(!existingEntity.getDate().equals(newEntity.getDate())) {
 			// different dates - unsupported
-		} else if(existingEntity.getTheaterCatalog().getTheaterMovieCatalogId().equals(newEntity.getTheaterCatalog().getTheaterMovieCatalogId())) {
+			throw new InvalidRequestException("Cannot update schedule date.");
+		} else if(!existingEntity.getTheaterCatalog().getTheaterMovieCatalogId().equals(newEntity.getTheaterCatalog().getTheaterMovieCatalogId())) {
 			// different theater and movie - unsupported
+			throw new InvalidRequestException("Cannot update theater or movie associated with this schedule");
 		} 
 
 		Map<ShowIdentifier, ShowEntity> existingShowMap = existingEntity.getShowList().parallelStream().collect(Collectors.toMap(ShowScheduleOpsHelper::getShowIdentifier, m->m));
@@ -89,10 +106,26 @@ public class ShowScheduleOpsHelper {
 			ShowEntity incomingShowEntity = entry.getValue();
 			if(existingShowMap.containsKey(entry.getKey()) ) {
 				ShowEntity existingShowEntity = existingShowMap.get(entry.getKey());
-				incomingShowEntity.getShowPricing().stream().forEach( p -> p.setShow(existingShowEntity));
+				
+				// compare and update pricing
+				Map<String, ShowPricingEntity> existingSeatPriceMap = existingShowEntity.getShowPricing().parallelStream().collect(Collectors.toMap(ShowPricingEntity::getSeatClass, m->m));
+				Map<String, ShowPricingEntity> newSeatPriceMap = incomingShowEntity.getShowPricing().parallelStream().collect(Collectors.toMap(ShowPricingEntity::getSeatClass, m->m));
+				List<ShowPricingEntity> toRemove = existingSeatPriceMap.entrySet().stream().filter(e->!newSeatPriceMap.containsKey(e.getKey())).map(Entry::getValue).collect(Collectors.toList());
+				List<ShowPricingEntity> toAdd = new ArrayList<>();
+				for(Entry<String, ShowPricingEntity> priceEntry : newSeatPriceMap.entrySet()) {
+					ShowPricingEntity existingPrice = existingSeatPriceMap.get(priceEntry.getKey());
+					if(existingPrice!=null) {
+						existingPrice.setPrice(priceEntry.getValue().getPrice());
+					} else {
+						ShowPricingEntity newPrice = priceEntry.getValue();
+						newPrice.setShow(existingShowEntity);
+						toAdd.add(priceEntry.getValue());
+					}
+				}
 				// update pricing
-				existingShowEntity.getShowPricing().clear();
-				existingShowEntity.getShowPricing().addAll(incomingShowEntity.getShowPricing());
+				existingShowEntity.getShowPricing().removeAll(toRemove);
+				existingShowEntity.getShowPricing().addAll(toAdd);
+				
 				// seat allocation for existing shows cannot be updated
 			} else {
 				newShowList.add(incomingShowEntity);
@@ -105,7 +138,14 @@ public class ShowScheduleOpsHelper {
 		existingEntity.getShowList().removeAll(showDeleteList);
 		
 		//save changes
-		ShowScheduleEntity updatedEntity = showScheduleRepository.saveAndFlush(existingEntity);
+		ShowScheduleEntity updatedEntity = null;
+		try {
+			updatedEntity = showScheduleRepository.saveAndFlush(existingEntity);
+		} catch(Exception e) {
+			String errorString = "Fail to update schedule. Exception: "+e.getMessage();
+			logger.error(errorString);
+			throw new PlatformServiceException(errorString, e);
+		}
 		
 		return showScheduleMapper.toRestObj(updatedEntity);
 	}
@@ -114,9 +154,19 @@ public class ShowScheduleOpsHelper {
 		return new ShowIdentifier(showEntity.getTime(), showEntity.getScreen().getScreenId());
 	}
 	
-	public void deleteShowSchedule(Long showScheduleId) {
+	public void deleteShowSchedule(Long showScheduleId) throws ResourceNotFoundException, PlatformServiceException {
 		ShowScheduleEntity showScheduleEntity = showScheduleRepository.lookupShowSchedule(showScheduleId);
-		showScheduleRepository.delete(showScheduleEntity);
+		if(showScheduleEntity==null) {
+			throw new ResourceNotFoundException("Fail to find schedule having id: "+showScheduleId);
+		}
+		
+		try {
+			showScheduleRepository.delete(showScheduleEntity);
+		} catch(Exception e) {
+			String errorString = "Fail to delete schedule. Exception: "+e.getMessage();
+			logger.error(errorString);
+			throw new PlatformServiceException(errorString, e);
+		}
 	}
 	
 	public static class ShowIdentifier {
